guidify-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic>=2.0
google-genai
supabase
PyPDF2
python-docx
pdf2image
pytesseract
python-multipart
pillow